+++ src/main/java/com/yt/mybatis/dao/BaseDaoAdapter.java
@@ -1,13 +1,12 @@
 package com.yt.mybatis.dao;
 
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-
 import com.yt.commons.Page;
 import com.yt.mybatis.model.BaseExample;
 import com.yt.mybatis.model.BaseMapper;
 import com.yt.mybatis.model.BaseModel;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import java.util.List;
 
 /**
  * Created by yt on 2017-7-6.
@@ -22,7 +21,7 @@ public abstract class BaseDaoAdapter<Model extends BaseModel, Example extends Ba
 	}
 
 	public Page<Model> queryOfPage(Example example) {
-		int cnt = getMapper().countByExample(example);
+		long cnt = getMapper().countByExample(example);
 
 		Page<Model> page = new Page<>();
 		page.setCurrentPage(example.getCurrentPage());
@@ -43,7 +42,7 @@ public abstract class BaseDaoAdapter<Model extends BaseModel, Example extends Ba
 	}
 
 	@Override
-	public int countByExample(Example example) {
+	public long countByExample(Example example) {
 		return getMapper().countByExample(example);
 	}
