+++ src/main/java/com/yt/mybatis/dao/BasePKDao.java
@@ -1,11 +1,11 @@
 package com.yt.mybatis.dao;
 
-import java.io.Serializable;
-import java.util.List;
-
 import com.yt.commons.Page;
 import com.yt.mybatis.model.BaseExample;
 import com.yt.mybatis.model.BaseModel;
+
+import java.io.Serializable;
+import java.util.List;
 
 /**
  * Created by yt on 2017-7-6.
@@ -15,7 +15,7 @@ public interface BasePKDao<PK extends Serializable, Model extends BaseModel, Exa
 
 	List<Model> selectPageByExample(Example example);
 
-	int countByExample(Example example);
+	long countByExample(Example example);
 
 	int deleteByExample(Example example);
